requirements.txt
# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20

# Data handling
requests>=2.31.0
pandas>=2.0.0

# Pydantic for structured outputs
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0
